Add move card.

+++ src/main/java/ru/feryafox/yetanotherkanbanboard/models/card/MoveCardDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class MoveCardDto {
     private Long newColumn;
-    private Long newPosition;
+    private Integer newPosition;
 }
